+++ csw-location/csw-location-api/jvm/src/main/scala/csw/location/api/AkkaRegistrationFactory.scala
@@ -1,28 +1,33 @@
 package csw.location.api
 
-import java.net.URI
-
+import akka.actor.typed.ActorRef
 import csw.location.api.commons.LocationServiceLogger
 import csw.location.api.exceptions.LocalAkkaActorRegistrationNotAllowed
+import csw.location.api.extensions.ActorExtension.RichActor
 import csw.location.api.models.Connection.AkkaConnection
 import csw.location.api.models.{AkkaRegistration, Metadata}
 import csw.logging.api.scaladsl.Logger
 
+object AkkaRegistrationFactory extends AkkaRegistrationFactory
+
+object JAkkaRegistrationFactory extends AkkaRegistrationFactory
+
 /**
  * A factory to create AkkaRegistration
  */
-object AkkaRegistrationFactory {
+class AkkaRegistrationFactory {
 
   /**
    *
     * @param connection the `Connection` to register with `LocationService`
-   * @param actorRefURI Provide a remote actor uri that is offering a connection. Local actors cannot be registered since they can't be
+   * @param actorRef Provide a remote actor ref that is offering a connection. Local actors cannot be registered since they can't be
    *                 communicated from components across the network
    * @param metadata represents additional information associated with registration
    * @return AkkaRegistration instance. A [[csw.location.api.exceptions.LocalAkkaActorRegistrationNotAllowed]]
    *        is thrown if the actorRefURI provided is not a remote actorRef uri
    */
-  def make(connection: AkkaConnection, actorRefURI: URI, metadata: Metadata): AkkaRegistration = {
+  def make(connection: AkkaConnection, actorRef: ActorRef[_], metadata: Metadata): AkkaRegistration = {
+    val actorRefURI = actorRef.toURI
     if (actorRefURI.getPort == -1) {
       val log: Logger            = LocationServiceLogger.getLogger
       val registrationNotAllowed = LocalAkkaActorRegistrationNotAllowed(actorRefURI)
@@ -36,12 +41,12 @@ object AkkaRegistrationFactory {
   /**
    *
     * @param connection the `Connection` to register with `LocationService`
-   * @param actorRefURI Provide a remote actor uri that is offering a connection. Local actors cannot be registered since they can't be
+   * @param actorRef Provide a remote actor ref that is offering a connection. Local actors cannot be registered since they can't be
    *                 communicated from components across the network
    * @return AkkaRegistration instance. A [[csw.location.api.exceptions.LocalAkkaActorRegistrationNotAllowed]]
    *        is thrown if the actorRefURI provided is not a remote actorRef uri
    */
-  def make(connection: AkkaConnection, actorRefURI: URI): AkkaRegistration =
-    make(connection, actorRefURI, Metadata.empty)
+  def make(connection: AkkaConnection, actorRef: ActorRef[_]): AkkaRegistration =
+    make(connection, actorRef, Metadata.empty)
 
 }